- Assets: Remplacement de quelques (int) field par des types réels (ObjectField). - Assets: Panneau ObjectEditorPageAmortissements avec un nouveau contenu.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2010.1@22058 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Products/Assets.App/Views/AbstractObjectEditorPage.cs
@@ -268,7 +268,7 @@ namespace Epsitec.Cresus.Assets.App.Views
 
 		protected void ShowHistoryPopup(Widget target, ObjectField field)
 		{
-			var popup = new HistoryPopup (this.accessor, this.baseType, this.objectGuid, this.timestamp, (int) field);
+			var popup = new HistoryPopup (this.accessor, this.baseType, this.objectGuid, this.timestamp, field);
 
 			popup.Create (target, leftOrRight: true);
